Refactor flop utils
