+++ src/game/aggregation.rs
@@ -7,8 +7,9 @@ use std::{
 };
 
 use crate::{
-    card_to_string, compute_average, game::utils::flop_helper::flop_to_string, Action, ActionTree,
-    Game, PostFlopGame, TreeConfig,
+    card_to_string, flop_to_string,
+    utils::stats::{get_action_evs, get_action_frequencies, get_player_stats},
+    Action, ActionTree, PostFlopGame, TreeConfig,
 };
 
 // TODO do we ever realistically want the Skip option?
@@ -23,76 +24,6 @@ pub enum ExistingReportBehavior {
     Overwrite,
     /// Describes the behavior of returning an error when an existing report is encountered.
     Error,
-}
-
-/// Returns the player's equity, EV, and EQR (in that order).
-/// *Requires game.cache_normalized_weights() to be called beforehand.*
-/// Equity and EQR are float values in the range [0.0, 1.0].
-/// Expected Value (EV) is a float representing the weighted average EV across all hands for the player at the current spot in game.
-///
-/// # Panics
-///
-/// A panic will occur if the input game is not solved.
-/// Additionally, a panic will occur if game.cache_normalized_weights() is not called before this function.
-fn get_player_stats(game: &PostFlopGame, player: usize) -> (f32, f32, f32) {
-    let equity = game.equity(player);
-    let ev = game.expected_values(player);
-    let weights = game.normalized_weights(player);
-    let average_equity = compute_average(&equity, weights);
-    let average_ev = compute_average(&ev, weights);
-    (
-        average_equity,
-        average_ev,
-        // Compute EQR
-        average_ev / (average_equity * game.pot() as f32),
-    )
-}
-
-/// Return the action frequencies of the current player.
-///
-/// # Panics
-///
-/// A panic will occur if the input game is not solved.
-/// Additionally, a panic will occur if game.cache_normalized_weights() is not called before this function.
-fn get_action_frequencies(game: &PostFlopGame) -> Vec<f32> {
-    let player = game.current_player();
-    let cards = game.private_cards(player);
-    let strategy = game.strategy();
-    let actions = game.available_actions();
-    let weights = game.normalized_weights(player);
-
-    (0..actions.len())
-        .map(|i| compute_average(&strategy[i * cards.len()..(i + 1) * cards.len()], weights))
-        .collect()
-}
-
-/// Returns the action EVs of the current player.
-///
-/// # Panics
-/// A panic will occur if the input game is not solved.
-/// Additionally, a panic will occur if game.cache_normalized_weights() is not called before this function.
-fn get_action_evs(game: &mut PostFlopGame) -> Vec<f32> {
-    let actions = game.available_actions();
-
-    (0..actions.len())
-        .map(|action_index| {
-            let player = game.current_player();
-            let num_private_hands = game.num_private_hands(player);
-            let relevant_range =
-                action_index * num_private_hands..(action_index + 1) * num_private_hands;
-
-            let strategy = &game.strategy()[relevant_range.clone()];
-            let evs_detail = &game.expected_values_detail(player)[relevant_range];
-            let norm_weights = game.normalized_weights(player);
-            let weights: Vec<f32> = strategy
-                .iter()
-                .zip(norm_weights)
-                .map(|(s, w)| s * w)
-                .collect();
-
-            compute_average(evs_detail, &weights)
-        })
-        .collect()
 }
 
 fn folder_name_from_action(action: Action) -> String {
@@ -539,6 +470,8 @@ pub fn load_test_game_and_config() -> (PostFlopGame, TreeConfig) {
 
 #[cfg(test)]
 mod tests {
+    use crate::compute_average;
+
     use super::*;
 
     fn get_current_player(prev_actions: &[Action], starting_player: usize) -> usize {
