--- src/game/utils/flop_helper.rs
@@ -1,262 +1,0 @@
-use std::collections::HashSet;
-
-use crate::{card_to_string, flop_from_str, Card};
-use lazy_static::lazy_static;
-
-lazy_static! {
-    static ref ALL_FLOPS: Vec<[Card; 3]> = [
-        "4c3c2c", "5c3c2c", "6c3c2c", "7c3c2c", "8c3c2c", "9c3c2c", "Tc3c2c", "Jc3c2c", "Qc3c2c",
-        "Kc3c2c", "Ac3c2c", "3c2c2d", "3c3d2c", "4d3c2c", "5d3c2c", "6d3c2c", "7d3c2c", "8d3c2c",
-        "9d3c2c", "Td3c2c", "Jd3c2c", "Qd3c2c", "Kd3c2c", "Ad3c2c", "5c4c2c", "6c4c2c", "7c4c2c",
-        "8c4c2c", "9c4c2c", "Tc4c2c", "Jc4c2c", "Qc4c2c", "Kc4c2c", "Ac4c2c", "4c2c2d", "4c3d2c",
-        "4c4d2c", "5d4c2c", "6d4c2c", "7d4c2c", "8d4c2c", "9d4c2c", "Td4c2c", "Jd4c2c", "Qd4c2c",
-        "Kd4c2c", "Ad4c2c", "6c5c2c", "7c5c2c", "8c5c2c", "9c5c2c", "Tc5c2c", "Jc5c2c", "Qc5c2c",
-        "Kc5c2c", "Ac5c2c", "5c2c2d", "5c3d2c", "5c4d2c", "5c5d2c", "6d5c2c", "7d5c2c", "8d5c2c",
-        "9d5c2c", "Td5c2c", "Jd5c2c", "Qd5c2c", "Kd5c2c", "Ad5c2c", "7c6c2c", "8c6c2c", "9c6c2c",
-        "Tc6c2c", "Jc6c2c", "Qc6c2c", "Kc6c2c", "Ac6c2c", "6c2c2d", "6c3d2c", "6c4d2c", "6c5d2c",
-        "6c6d2c", "7d6c2c", "8d6c2c", "9d6c2c", "Td6c2c", "Jd6c2c", "Qd6c2c", "Kd6c2c", "Ad6c2c",
-        "8c7c2c", "9c7c2c", "Tc7c2c", "Jc7c2c", "Qc7c2c", "Kc7c2c", "Ac7c2c", "7c2c2d", "7c3d2c",
-        "7c4d2c", "7c5d2c", "7c6d2c", "7c7d2c", "8d7c2c", "9d7c2c", "Td7c2c", "Jd7c2c", "Qd7c2c",
-        "Kd7c2c", "Ad7c2c", "9c8c2c", "Tc8c2c", "Jc8c2c", "Qc8c2c", "Kc8c2c", "Ac8c2c", "8c2c2d",
-        "8c3d2c", "8c4d2c", "8c5d2c", "8c6d2c", "8c7d2c", "8c8d2c", "9d8c2c", "Td8c2c", "Jd8c2c",
-        "Qd8c2c", "Kd8c2c", "Ad8c2c", "Tc9c2c", "Jc9c2c", "Qc9c2c", "Kc9c2c", "Ac9c2c", "9c2c2d",
-        "9c3d2c", "9c4d2c", "9c5d2c", "9c6d2c", "9c7d2c", "9c8d2c", "9c9d2c", "Td9c2c", "Jd9c2c",
-        "Qd9c2c", "Kd9c2c", "Ad9c2c", "JcTc2c", "QcTc2c", "KcTc2c", "AcTc2c", "Tc2c2d", "Tc3d2c",
-        "Tc4d2c", "Tc5d2c", "Tc6d2c", "Tc7d2c", "Tc8d2c", "Tc9d2c", "TcTd2c", "JdTc2c", "QdTc2c",
-        "KdTc2c", "AdTc2c", "QcJc2c", "KcJc2c", "AcJc2c", "Jc2c2d", "Jc3d2c", "Jc4d2c", "Jc5d2c",
-        "Jc6d2c", "Jc7d2c", "Jc8d2c", "Jc9d2c", "JcTd2c", "JcJd2c", "QdJc2c", "KdJc2c", "AdJc2c",
-        "KcQc2c", "AcQc2c", "Qc2c2d", "Qc3d2c", "Qc4d2c", "Qc5d2c", "Qc6d2c", "Qc7d2c", "Qc8d2c",
-        "Qc9d2c", "QcTd2c", "QcJd2c", "QcQd2c", "KdQc2c", "AdQc2c", "AcKc2c", "Kc2c2d", "Kc3d2c",
-        "Kc4d2c", "Kc5d2c", "Kc6d2c", "Kc7d2c", "Kc8d2c", "Kc9d2c", "KcTd2c", "KcJd2c", "KcQd2c",
-        "KcKd2c", "AdKc2c", "Ac2c2d", "Ac3d2c", "Ac4d2c", "Ac5d2c", "Ac6d2c", "Ac7d2c", "Ac8d2c",
-        "Ac9d2c", "AcTd2c", "AcJd2c", "AcQd2c", "AcKd2c", "AcAd2c", "2c2d2h", "3h2c2d", "4h2c2d",
-        "5h2c2d", "6h2c2d", "7h2c2d", "8h2c2d", "9h2c2d", "Th2c2d", "Jh2c2d", "Qh2c2d", "Kh2c2d",
-        "Ah2c2d", "3d3h2c", "4h3d2c", "5h3d2c", "6h3d2c", "7h3d2c", "8h3d2c", "9h3d2c", "Th3d2c",
-        "Jh3d2c", "Qh3d2c", "Kh3d2c", "Ah3d2c", "4d4h2c", "5h4d2c", "6h4d2c", "7h4d2c", "8h4d2c",
-        "9h4d2c", "Th4d2c", "Jh4d2c", "Qh4d2c", "Kh4d2c", "Ah4d2c", "5d5h2c", "6h5d2c", "7h5d2c",
-        "8h5d2c", "9h5d2c", "Th5d2c", "Jh5d2c", "Qh5d2c", "Kh5d2c", "Ah5d2c", "6d6h2c", "7h6d2c",
-        "8h6d2c", "9h6d2c", "Th6d2c", "Jh6d2c", "Qh6d2c", "Kh6d2c", "Ah6d2c", "7d7h2c", "8h7d2c",
-        "9h7d2c", "Th7d2c", "Jh7d2c", "Qh7d2c", "Kh7d2c", "Ah7d2c", "8d8h2c", "9h8d2c", "Th8d2c",
-        "Jh8d2c", "Qh8d2c", "Kh8d2c", "Ah8d2c", "9d9h2c", "Th9d2c", "Jh9d2c", "Qh9d2c", "Kh9d2c",
-        "Ah9d2c", "TdTh2c", "JhTd2c", "QhTd2c", "KhTd2c", "AhTd2c", "JdJh2c", "QhJd2c", "KhJd2c",
-        "AhJd2c", "QdQh2c", "KhQd2c", "AhQd2c", "KdKh2c", "AhKd2c", "AdAh2c", "5c4c3c", "6c4c3c",
-        "7c4c3c", "8c4c3c", "9c4c3c", "Tc4c3c", "Jc4c3c", "Qc4c3c", "Kc4c3c", "Ac4c3c", "4c3c2d",
-        "4c3c3d", "4c4d3c", "5d4c3c", "6d4c3c", "7d4c3c", "8d4c3c", "9d4c3c", "Td4c3c", "Jd4c3c",
-        "Qd4c3c", "Kd4c3c", "Ad4c3c", "6c5c3c", "7c5c3c", "8c5c3c", "9c5c3c", "Tc5c3c", "Jc5c3c",
-        "Qc5c3c", "Kc5c3c", "Ac5c3c", "5c3c2d", "5c3c3d", "5c4d3c", "5c5d3c", "6d5c3c", "7d5c3c",
-        "8d5c3c", "9d5c3c", "Td5c3c", "Jd5c3c", "Qd5c3c", "Kd5c3c", "Ad5c3c", "7c6c3c", "8c6c3c",
-        "9c6c3c", "Tc6c3c", "Jc6c3c", "Qc6c3c", "Kc6c3c", "Ac6c3c", "6c3c2d", "6c3c3d", "6c4d3c",
-        "6c5d3c", "6c6d3c", "7d6c3c", "8d6c3c", "9d6c3c", "Td6c3c", "Jd6c3c", "Qd6c3c", "Kd6c3c",
-        "Ad6c3c", "8c7c3c", "9c7c3c", "Tc7c3c", "Jc7c3c", "Qc7c3c", "Kc7c3c", "Ac7c3c", "7c3c2d",
-        "7c3c3d", "7c4d3c", "7c5d3c", "7c6d3c", "7c7d3c", "8d7c3c", "9d7c3c", "Td7c3c", "Jd7c3c",
-        "Qd7c3c", "Kd7c3c", "Ad7c3c", "9c8c3c", "Tc8c3c", "Jc8c3c", "Qc8c3c", "Kc8c3c", "Ac8c3c",
-        "8c3c2d", "8c3c3d", "8c4d3c", "8c5d3c", "8c6d3c", "8c7d3c", "8c8d3c", "9d8c3c", "Td8c3c",
-        "Jd8c3c", "Qd8c3c", "Kd8c3c", "Ad8c3c", "Tc9c3c", "Jc9c3c", "Qc9c3c", "Kc9c3c", "Ac9c3c",
-        "9c3c2d", "9c3c3d", "9c4d3c", "9c5d3c", "9c6d3c", "9c7d3c", "9c8d3c", "9c9d3c", "Td9c3c",
-        "Jd9c3c", "Qd9c3c", "Kd9c3c", "Ad9c3c", "JcTc3c", "QcTc3c", "KcTc3c", "AcTc3c", "Tc3c2d",
-        "Tc3c3d", "Tc4d3c", "Tc5d3c", "Tc6d3c", "Tc7d3c", "Tc8d3c", "Tc9d3c", "TcTd3c", "JdTc3c",
-        "QdTc3c", "KdTc3c", "AdTc3c", "QcJc3c", "KcJc3c", "AcJc3c", "Jc3c2d", "Jc3c3d", "Jc4d3c",
-        "Jc5d3c", "Jc6d3c", "Jc7d3c", "Jc8d3c", "Jc9d3c", "JcTd3c", "JcJd3c", "QdJc3c", "KdJc3c",
-        "AdJc3c", "KcQc3c", "AcQc3c", "Qc3c2d", "Qc3c3d", "Qc4d3c", "Qc5d3c", "Qc6d3c", "Qc7d3c",
-        "Qc8d3c", "Qc9d3c", "QcTd3c", "QcJd3c", "QcQd3c", "KdQc3c", "AdQc3c", "AcKc3c", "Kc3c2d",
-        "Kc3c3d", "Kc4d3c", "Kc5d3c", "Kc6d3c", "Kc7d3c", "Kc8d3c", "Kc9d3c", "KcTd3c", "KcJd3c",
-        "KcQd3c", "KcKd3c", "AdKc3c", "Ac3c2d", "Ac3c3d", "Ac4d3c", "Ac5d3c", "Ac6d3c", "Ac7d3c",
-        "Ac8d3c", "Ac9d3c", "AcTd3c", "AcJd3c", "AcQd3c", "AcKd3c", "AcAd3c", "3c3d3h", "4h3c3d",
-        "5h3c3d", "6h3c3d", "7h3c3d", "8h3c3d", "9h3c3d", "Th3c3d", "Jh3c3d", "Qh3c3d", "Kh3c3d",
-        "Ah3c3d", "4d4h3c", "5h4d3c", "6h4d3c", "7h4d3c", "8h4d3c", "9h4d3c", "Th4d3c", "Jh4d3c",
-        "Qh4d3c", "Kh4d3c", "Ah4d3c", "5d5h3c", "6h5d3c", "7h5d3c", "8h5d3c", "9h5d3c", "Th5d3c",
-        "Jh5d3c", "Qh5d3c", "Kh5d3c", "Ah5d3c", "6d6h3c", "7h6d3c", "8h6d3c", "9h6d3c", "Th6d3c",
-        "Jh6d3c", "Qh6d3c", "Kh6d3c", "Ah6d3c", "7d7h3c", "8h7d3c", "9h7d3c", "Th7d3c", "Jh7d3c",
-        "Qh7d3c", "Kh7d3c", "Ah7d3c", "8d8h3c", "9h8d3c", "Th8d3c", "Jh8d3c", "Qh8d3c", "Kh8d3c",
-        "Ah8d3c", "9d9h3c", "Th9d3c", "Jh9d3c", "Qh9d3c", "Kh9d3c", "Ah9d3c", "TdTh3c", "JhTd3c",
-        "QhTd3c", "KhTd3c", "AhTd3c", "JdJh3c", "QhJd3c", "KhJd3c", "AhJd3c", "QdQh3c", "KhQd3c",
-        "AhQd3c", "KdKh3c", "AhKd3c", "AdAh3c", "6c5c4c", "7c5c4c", "8c5c4c", "9c5c4c", "Tc5c4c",
-        "Jc5c4c", "Qc5c4c", "Kc5c4c", "Ac5c4c", "5c4c2d", "5c4c3d", "5c4c4d", "5c5d4c", "6d5c4c",
-        "7d5c4c", "8d5c4c", "9d5c4c", "Td5c4c", "Jd5c4c", "Qd5c4c", "Kd5c4c", "Ad5c4c", "7c6c4c",
-        "8c6c4c", "9c6c4c", "Tc6c4c", "Jc6c4c", "Qc6c4c", "Kc6c4c", "Ac6c4c", "6c4c2d", "6c4c3d",
-        "6c4c4d", "6c5d4c", "6c6d4c", "7d6c4c", "8d6c4c", "9d6c4c", "Td6c4c", "Jd6c4c", "Qd6c4c",
-        "Kd6c4c", "Ad6c4c", "8c7c4c", "9c7c4c", "Tc7c4c", "Jc7c4c", "Qc7c4c", "Kc7c4c", "Ac7c4c",
-        "7c4c2d", "7c4c3d", "7c4c4d", "7c5d4c", "7c6d4c", "7c7d4c", "8d7c4c", "9d7c4c", "Td7c4c",
-        "Jd7c4c", "Qd7c4c", "Kd7c4c", "Ad7c4c", "9c8c4c", "Tc8c4c", "Jc8c4c", "Qc8c4c", "Kc8c4c",
-        "Ac8c4c", "8c4c2d", "8c4c3d", "8c4c4d", "8c5d4c", "8c6d4c", "8c7d4c", "8c8d4c", "9d8c4c",
-        "Td8c4c", "Jd8c4c", "Qd8c4c", "Kd8c4c", "Ad8c4c", "Tc9c4c", "Jc9c4c", "Qc9c4c", "Kc9c4c",
-        "Ac9c4c", "9c4c2d", "9c4c3d", "9c4c4d", "9c5d4c", "9c6d4c", "9c7d4c", "9c8d4c", "9c9d4c",
-        "Td9c4c", "Jd9c4c", "Qd9c4c", "Kd9c4c", "Ad9c4c", "JcTc4c", "QcTc4c", "KcTc4c", "AcTc4c",
-        "Tc4c2d", "Tc4c3d", "Tc4c4d", "Tc5d4c", "Tc6d4c", "Tc7d4c", "Tc8d4c", "Tc9d4c", "TcTd4c",
-        "JdTc4c", "QdTc4c", "KdTc4c", "AdTc4c", "QcJc4c", "KcJc4c", "AcJc4c", "Jc4c2d", "Jc4c3d",
-        "Jc4c4d", "Jc5d4c", "Jc6d4c", "Jc7d4c", "Jc8d4c", "Jc9d4c", "JcTd4c", "JcJd4c", "QdJc4c",
-        "KdJc4c", "AdJc4c", "KcQc4c", "AcQc4c", "Qc4c2d", "Qc4c3d", "Qc4c4d", "Qc5d4c", "Qc6d4c",
-        "Qc7d4c", "Qc8d4c", "Qc9d4c", "QcTd4c", "QcJd4c", "QcQd4c", "KdQc4c", "AdQc4c", "AcKc4c",
-        "Kc4c2d", "Kc4c3d", "Kc4c4d", "Kc5d4c", "Kc6d4c", "Kc7d4c", "Kc8d4c", "Kc9d4c", "KcTd4c",
-        "KcJd4c", "KcQd4c", "KcKd4c", "AdKc4c", "Ac4c2d", "Ac4c3d", "Ac4c4d", "Ac5d4c", "Ac6d4c",
-        "Ac7d4c", "Ac8d4c", "Ac9d4c", "AcTd4c", "AcJd4c", "AcQd4c", "AcKd4c", "AcAd4c", "4c4d4h",
-        "5h4c4d", "6h4c4d", "7h4c4d", "8h4c4d", "9h4c4d", "Th4c4d", "Jh4c4d", "Qh4c4d", "Kh4c4d",
-        "Ah4c4d", "5d5h4c", "6h5d4c", "7h5d4c", "8h5d4c", "9h5d4c", "Th5d4c", "Jh5d4c", "Qh5d4c",
-        "Kh5d4c", "Ah5d4c", "6d6h4c", "7h6d4c", "8h6d4c", "9h6d4c", "Th6d4c", "Jh6d4c", "Qh6d4c",
-        "Kh6d4c", "Ah6d4c", "7d7h4c", "8h7d4c", "9h7d4c", "Th7d4c", "Jh7d4c", "Qh7d4c", "Kh7d4c",
-        "Ah7d4c", "8d8h4c", "9h8d4c", "Th8d4c", "Jh8d4c", "Qh8d4c", "Kh8d4c", "Ah8d4c", "9d9h4c",
-        "Th9d4c", "Jh9d4c", "Qh9d4c", "Kh9d4c", "Ah9d4c", "TdTh4c", "JhTd4c", "QhTd4c", "KhTd4c",
-        "AhTd4c", "JdJh4c", "QhJd4c", "KhJd4c", "AhJd4c", "QdQh4c", "KhQd4c", "AhQd4c", "KdKh4c",
-        "AhKd4c", "AdAh4c", "7c6c5c", "8c6c5c", "9c6c5c", "Tc6c5c", "Jc6c5c", "Qc6c5c", "Kc6c5c",
-        "Ac6c5c", "6c5c2d", "6c5c3d", "6c5c4d", "6c5c5d", "6c6d5c", "7d6c5c", "8d6c5c", "9d6c5c",
-        "Td6c5c", "Jd6c5c", "Qd6c5c", "Kd6c5c", "Ad6c5c", "8c7c5c", "9c7c5c", "Tc7c5c", "Jc7c5c",
-        "Qc7c5c", "Kc7c5c", "Ac7c5c", "7c5c2d", "7c5c3d", "7c5c4d", "7c5c5d", "7c6d5c", "7c7d5c",
-        "8d7c5c", "9d7c5c", "Td7c5c", "Jd7c5c", "Qd7c5c", "Kd7c5c", "Ad7c5c", "9c8c5c", "Tc8c5c",
-        "Jc8c5c", "Qc8c5c", "Kc8c5c", "Ac8c5c", "8c5c2d", "8c5c3d", "8c5c4d", "8c5c5d", "8c6d5c",
-        "8c7d5c", "8c8d5c", "9d8c5c", "Td8c5c", "Jd8c5c", "Qd8c5c", "Kd8c5c", "Ad8c5c", "Tc9c5c",
-        "Jc9c5c", "Qc9c5c", "Kc9c5c", "Ac9c5c", "9c5c2d", "9c5c3d", "9c5c4d", "9c5c5d", "9c6d5c",
-        "9c7d5c", "9c8d5c", "9c9d5c", "Td9c5c", "Jd9c5c", "Qd9c5c", "Kd9c5c", "Ad9c5c", "JcTc5c",
-        "QcTc5c", "KcTc5c", "AcTc5c", "Tc5c2d", "Tc5c3d", "Tc5c4d", "Tc5c5d", "Tc6d5c", "Tc7d5c",
-        "Tc8d5c", "Tc9d5c", "TcTd5c", "JdTc5c", "QdTc5c", "KdTc5c", "AdTc5c", "QcJc5c", "KcJc5c",
-        "AcJc5c", "Jc5c2d", "Jc5c3d", "Jc5c4d", "Jc5c5d", "Jc6d5c", "Jc7d5c", "Jc8d5c", "Jc9d5c",
-        "JcTd5c", "JcJd5c", "QdJc5c", "KdJc5c", "AdJc5c", "KcQc5c", "AcQc5c", "Qc5c2d", "Qc5c3d",
-        "Qc5c4d", "Qc5c5d", "Qc6d5c", "Qc7d5c", "Qc8d5c", "Qc9d5c", "QcTd5c", "QcJd5c", "QcQd5c",
-        "KdQc5c", "AdQc5c", "AcKc5c", "Kc5c2d", "Kc5c3d", "Kc5c4d", "Kc5c5d", "Kc6d5c", "Kc7d5c",
-        "Kc8d5c", "Kc9d5c", "KcTd5c", "KcJd5c", "KcQd5c", "KcKd5c", "AdKc5c", "Ac5c2d", "Ac5c3d",
-        "Ac5c4d", "Ac5c5d", "Ac6d5c", "Ac7d5c", "Ac8d5c", "Ac9d5c", "AcTd5c", "AcJd5c", "AcQd5c",
-        "AcKd5c", "AcAd5c", "5c5d5h", "6h5c5d", "7h5c5d", "8h5c5d", "9h5c5d", "Th5c5d", "Jh5c5d",
-        "Qh5c5d", "Kh5c5d", "Ah5c5d", "6d6h5c", "7h6d5c", "8h6d5c", "9h6d5c", "Th6d5c", "Jh6d5c",
-        "Qh6d5c", "Kh6d5c", "Ah6d5c", "7d7h5c", "8h7d5c", "9h7d5c", "Th7d5c", "Jh7d5c", "Qh7d5c",
-        "Kh7d5c", "Ah7d5c", "8d8h5c", "9h8d5c", "Th8d5c", "Jh8d5c", "Qh8d5c", "Kh8d5c", "Ah8d5c",
-        "9d9h5c", "Th9d5c", "Jh9d5c", "Qh9d5c", "Kh9d5c", "Ah9d5c", "TdTh5c", "JhTd5c", "QhTd5c",
-        "KhTd5c", "AhTd5c", "JdJh5c", "QhJd5c", "KhJd5c", "AhJd5c", "QdQh5c", "KhQd5c", "AhQd5c",
-        "KdKh5c", "AhKd5c", "AdAh5c", "8c7c6c", "9c7c6c", "Tc7c6c", "Jc7c6c", "Qc7c6c", "Kc7c6c",
-        "Ac7c6c", "7c6c2d", "7c6c3d", "7c6c4d", "7c6c5d", "7c6c6d", "7c7d6c", "8d7c6c", "9d7c6c",
-        "Td7c6c", "Jd7c6c", "Qd7c6c", "Kd7c6c", "Ad7c6c", "9c8c6c", "Tc8c6c", "Jc8c6c", "Qc8c6c",
-        "Kc8c6c", "Ac8c6c", "8c6c2d", "8c6c3d", "8c6c4d", "8c6c5d", "8c6c6d", "8c7d6c", "8c8d6c",
-        "9d8c6c", "Td8c6c", "Jd8c6c", "Qd8c6c", "Kd8c6c", "Ad8c6c", "Tc9c6c", "Jc9c6c", "Qc9c6c",
-        "Kc9c6c", "Ac9c6c", "9c6c2d", "9c6c3d", "9c6c4d", "9c6c5d", "9c6c6d", "9c7d6c", "9c8d6c",
-        "9c9d6c", "Td9c6c", "Jd9c6c", "Qd9c6c", "Kd9c6c", "Ad9c6c", "JcTc6c", "QcTc6c", "KcTc6c",
-        "AcTc6c", "Tc6c2d", "Tc6c3d", "Tc6c4d", "Tc6c5d", "Tc6c6d", "Tc7d6c", "Tc8d6c", "Tc9d6c",
-        "TcTd6c", "JdTc6c", "QdTc6c", "KdTc6c", "AdTc6c", "QcJc6c", "KcJc6c", "AcJc6c", "Jc6c2d",
-        "Jc6c3d", "Jc6c4d", "Jc6c5d", "Jc6c6d", "Jc7d6c", "Jc8d6c", "Jc9d6c", "JcTd6c", "JcJd6c",
-        "QdJc6c", "KdJc6c", "AdJc6c", "KcQc6c", "AcQc6c", "Qc6c2d", "Qc6c3d", "Qc6c4d", "Qc6c5d",
-        "Qc6c6d", "Qc7d6c", "Qc8d6c", "Qc9d6c", "QcTd6c", "QcJd6c", "QcQd6c", "KdQc6c", "AdQc6c",
-        "AcKc6c", "Kc6c2d", "Kc6c3d", "Kc6c4d", "Kc6c5d", "Kc6c6d", "Kc7d6c", "Kc8d6c", "Kc9d6c",
-        "KcTd6c", "KcJd6c", "KcQd6c", "KcKd6c", "AdKc6c", "Ac6c2d", "Ac6c3d", "Ac6c4d", "Ac6c5d",
-        "Ac6c6d", "Ac7d6c", "Ac8d6c", "Ac9d6c", "AcTd6c", "AcJd6c", "AcQd6c", "AcKd6c", "AcAd6c",
-        "6c6d6h", "7h6c6d", "8h6c6d", "9h6c6d", "Th6c6d", "Jh6c6d", "Qh6c6d", "Kh6c6d", "Ah6c6d",
-        "7d7h6c", "8h7d6c", "9h7d6c", "Th7d6c", "Jh7d6c", "Qh7d6c", "Kh7d6c", "Ah7d6c", "8d8h6c",
-        "9h8d6c", "Th8d6c", "Jh8d6c", "Qh8d6c", "Kh8d6c", "Ah8d6c", "9d9h6c", "Th9d6c", "Jh9d6c",
-        "Qh9d6c", "Kh9d6c", "Ah9d6c", "TdTh6c", "JhTd6c", "QhTd6c", "KhTd6c", "AhTd6c", "JdJh6c",
-        "QhJd6c", "KhJd6c", "AhJd6c", "QdQh6c", "KhQd6c", "AhQd6c", "KdKh6c", "AhKd6c", "AdAh6c",
-        "9c8c7c", "Tc8c7c", "Jc8c7c", "Qc8c7c", "Kc8c7c", "Ac8c7c", "8c7c2d", "8c7c3d", "8c7c4d",
-        "8c7c5d", "8c7c6d", "8c7c7d", "8c8d7c", "9d8c7c", "Td8c7c", "Jd8c7c", "Qd8c7c", "Kd8c7c",
-        "Ad8c7c", "Tc9c7c", "Jc9c7c", "Qc9c7c", "Kc9c7c", "Ac9c7c", "9c7c2d", "9c7c3d", "9c7c4d",
-        "9c7c5d", "9c7c6d", "9c7c7d", "9c8d7c", "9c9d7c", "Td9c7c", "Jd9c7c", "Qd9c7c", "Kd9c7c",
-        "Ad9c7c", "JcTc7c", "QcTc7c", "KcTc7c", "AcTc7c", "Tc7c2d", "Tc7c3d", "Tc7c4d", "Tc7c5d",
-        "Tc7c6d", "Tc7c7d", "Tc8d7c", "Tc9d7c", "TcTd7c", "JdTc7c", "QdTc7c", "KdTc7c", "AdTc7c",
-        "QcJc7c", "KcJc7c", "AcJc7c", "Jc7c2d", "Jc7c3d", "Jc7c4d", "Jc7c5d", "Jc7c6d", "Jc7c7d",
-        "Jc8d7c", "Jc9d7c", "JcTd7c", "JcJd7c", "QdJc7c", "KdJc7c", "AdJc7c", "KcQc7c", "AcQc7c",
-        "Qc7c2d", "Qc7c3d", "Qc7c4d", "Qc7c5d", "Qc7c6d", "Qc7c7d", "Qc8d7c", "Qc9d7c", "QcTd7c",
-        "QcJd7c", "QcQd7c", "KdQc7c", "AdQc7c", "AcKc7c", "Kc7c2d", "Kc7c3d", "Kc7c4d", "Kc7c5d",
-        "Kc7c6d", "Kc7c7d", "Kc8d7c", "Kc9d7c", "KcTd7c", "KcJd7c", "KcQd7c", "KcKd7c", "AdKc7c",
-        "Ac7c2d", "Ac7c3d", "Ac7c4d", "Ac7c5d", "Ac7c6d", "Ac7c7d", "Ac8d7c", "Ac9d7c", "AcTd7c",
-        "AcJd7c", "AcQd7c", "AcKd7c", "AcAd7c", "7c7d7h", "8h7c7d", "9h7c7d", "Th7c7d", "Jh7c7d",
-        "Qh7c7d", "Kh7c7d", "Ah7c7d", "8d8h7c", "9h8d7c", "Th8d7c", "Jh8d7c", "Qh8d7c", "Kh8d7c",
-        "Ah8d7c", "9d9h7c", "Th9d7c", "Jh9d7c", "Qh9d7c", "Kh9d7c", "Ah9d7c", "TdTh7c", "JhTd7c",
-        "QhTd7c", "KhTd7c", "AhTd7c", "JdJh7c", "QhJd7c", "KhJd7c", "AhJd7c", "QdQh7c", "KhQd7c",
-        "AhQd7c", "KdKh7c", "AhKd7c", "AdAh7c", "Tc9c8c", "Jc9c8c", "Qc9c8c", "Kc9c8c", "Ac9c8c",
-        "9c8c2d", "9c8c3d", "9c8c4d", "9c8c5d", "9c8c6d", "9c8c7d", "9c8c8d", "9c9d8c", "Td9c8c",
-        "Jd9c8c", "Qd9c8c", "Kd9c8c", "Ad9c8c", "JcTc8c", "QcTc8c", "KcTc8c", "AcTc8c", "Tc8c2d",
-        "Tc8c3d", "Tc8c4d", "Tc8c5d", "Tc8c6d", "Tc8c7d", "Tc8c8d", "Tc9d8c", "TcTd8c", "JdTc8c",
-        "QdTc8c", "KdTc8c", "AdTc8c", "QcJc8c", "KcJc8c", "AcJc8c", "Jc8c2d", "Jc8c3d", "Jc8c4d",
-        "Jc8c5d", "Jc8c6d", "Jc8c7d", "Jc8c8d", "Jc9d8c", "JcTd8c", "JcJd8c", "QdJc8c", "KdJc8c",
-        "AdJc8c", "KcQc8c", "AcQc8c", "Qc8c2d", "Qc8c3d", "Qc8c4d", "Qc8c5d", "Qc8c6d", "Qc8c7d",
-        "Qc8c8d", "Qc9d8c", "QcTd8c", "QcJd8c", "QcQd8c", "KdQc8c", "AdQc8c", "AcKc8c", "Kc8c2d",
-        "Kc8c3d", "Kc8c4d", "Kc8c5d", "Kc8c6d", "Kc8c7d", "Kc8c8d", "Kc9d8c", "KcTd8c", "KcJd8c",
-        "KcQd8c", "KcKd8c", "AdKc8c", "Ac8c2d", "Ac8c3d", "Ac8c4d", "Ac8c5d", "Ac8c6d", "Ac8c7d",
-        "Ac8c8d", "Ac9d8c", "AcTd8c", "AcJd8c", "AcQd8c", "AcKd8c", "AcAd8c", "8c8d8h", "9h8c8d",
-        "Th8c8d", "Jh8c8d", "Qh8c8d", "Kh8c8d", "Ah8c8d", "9d9h8c", "Th9d8c", "Jh9d8c", "Qh9d8c",
-        "Kh9d8c", "Ah9d8c", "TdTh8c", "JhTd8c", "QhTd8c", "KhTd8c", "AhTd8c", "JdJh8c", "QhJd8c",
-        "KhJd8c", "AhJd8c", "QdQh8c", "KhQd8c", "AhQd8c", "KdKh8c", "AhKd8c", "AdAh8c", "JcTc9c",
-        "QcTc9c", "KcTc9c", "AcTc9c", "Tc9c2d", "Tc9c3d", "Tc9c4d", "Tc9c5d", "Tc9c6d", "Tc9c7d",
-        "Tc9c8d", "Tc9c9d", "TcTd9c", "JdTc9c", "QdTc9c", "KdTc9c", "AdTc9c", "QcJc9c", "KcJc9c",
-        "AcJc9c", "Jc9c2d", "Jc9c3d", "Jc9c4d", "Jc9c5d", "Jc9c6d", "Jc9c7d", "Jc9c8d", "Jc9c9d",
-        "JcTd9c", "JcJd9c", "QdJc9c", "KdJc9c", "AdJc9c", "KcQc9c", "AcQc9c", "Qc9c2d", "Qc9c3d",
-        "Qc9c4d", "Qc9c5d", "Qc9c6d", "Qc9c7d", "Qc9c8d", "Qc9c9d", "QcTd9c", "QcJd9c", "QcQd9c",
-        "KdQc9c", "AdQc9c", "AcKc9c", "Kc9c2d", "Kc9c3d", "Kc9c4d", "Kc9c5d", "Kc9c6d", "Kc9c7d",
-        "Kc9c8d", "Kc9c9d", "KcTd9c", "KcJd9c", "KcQd9c", "KcKd9c", "AdKc9c", "Ac9c2d", "Ac9c3d",
-        "Ac9c4d", "Ac9c5d", "Ac9c6d", "Ac9c7d", "Ac9c8d", "Ac9c9d", "AcTd9c", "AcJd9c", "AcQd9c",
-        "AcKd9c", "AcAd9c", "9c9d9h", "Th9c9d", "Jh9c9d", "Qh9c9d", "Kh9c9d", "Ah9c9d", "TdTh9c",
-        "JhTd9c", "QhTd9c", "KhTd9c", "AhTd9c", "JdJh9c", "QhJd9c", "KhJd9c", "AhJd9c", "QdQh9c",
-        "KhQd9c", "AhQd9c", "KdKh9c", "AhKd9c", "AdAh9c", "QcJcTc", "KcJcTc", "AcJcTc", "JcTc2d",
-        "JcTc3d", "JcTc4d", "JcTc5d", "JcTc6d", "JcTc7d", "JcTc8d", "JcTc9d", "JcTcTd", "JcJdTc",
-        "QdJcTc", "KdJcTc", "AdJcTc", "KcQcTc", "AcQcTc", "QcTc2d", "QcTc3d", "QcTc4d", "QcTc5d",
-        "QcTc6d", "QcTc7d", "QcTc8d", "QcTc9d", "QcTcTd", "QcJdTc", "QcQdTc", "KdQcTc", "AdQcTc",
-        "AcKcTc", "KcTc2d", "KcTc3d", "KcTc4d", "KcTc5d", "KcTc6d", "KcTc7d", "KcTc8d", "KcTc9d",
-        "KcTcTd", "KcJdTc", "KcQdTc", "KcKdTc", "AdKcTc", "AcTc2d", "AcTc3d", "AcTc4d", "AcTc5d",
-        "AcTc6d", "AcTc7d", "AcTc8d", "AcTc9d", "AcTcTd", "AcJdTc", "AcQdTc", "AcKdTc", "AcAdTc",
-        "TcTdTh", "JhTcTd", "QhTcTd", "KhTcTd", "AhTcTd", "JdJhTc", "QhJdTc", "KhJdTc", "AhJdTc",
-        "QdQhTc", "KhQdTc", "AhQdTc", "KdKhTc", "AhKdTc", "AdAhTc", "KcQcJc", "AcQcJc", "QcJc2d",
-        "QcJc3d", "QcJc4d", "QcJc5d", "QcJc6d", "QcJc7d", "QcJc8d", "QcJc9d", "QcJcTd", "QcJcJd",
-        "QcQdJc", "KdQcJc", "AdQcJc", "AcKcJc", "KcJc2d", "KcJc3d", "KcJc4d", "KcJc5d", "KcJc6d",
-        "KcJc7d", "KcJc8d", "KcJc9d", "KcJcTd", "KcJcJd", "KcQdJc", "KcKdJc", "AdKcJc", "AcJc2d",
-        "AcJc3d", "AcJc4d", "AcJc5d", "AcJc6d", "AcJc7d", "AcJc8d", "AcJc9d", "AcJcTd", "AcJcJd",
-        "AcQdJc", "AcKdJc", "AcAdJc", "JcJdJh", "QhJcJd", "KhJcJd", "AhJcJd", "QdQhJc", "KhQdJc",
-        "AhQdJc", "KdKhJc", "AhKdJc", "AdAhJc", "AcKcQc", "KcQc2d", "KcQc3d", "KcQc4d", "KcQc5d",
-        "KcQc6d", "KcQc7d", "KcQc8d", "KcQc9d", "KcQcTd", "KcQcJd", "KcQcQd", "KcKdQc", "AdKcQc",
-        "AcQc2d", "AcQc3d", "AcQc4d", "AcQc5d", "AcQc6d", "AcQc7d", "AcQc8d", "AcQc9d", "AcQcTd",
-        "AcQcJd", "AcQcQd", "AcKdQc", "AcAdQc", "QcQdQh", "KhQcQd", "AhQcQd", "KdKhQc", "AhKdQc",
-        "AdAhQc", "AcKc2d", "AcKc3d", "AcKc4d", "AcKc5d", "AcKc6d", "AcKc7d", "AcKc8d", "AcKc9d",
-        "AcKcTd", "AcKcJd", "AcKcQd", "AcKcKd", "AcAdKc", "KcKdKh", "AhKcKd", "AdAhKc", "AcAdAh",
-    ]
-    .iter()
-    .map(|flop_str| flop_from_str(flop_str).unwrap())
-    .collect();
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-pub enum Texture {
-    Monotone,
-    Twotone,
-    Rainbow,
-    Paired,
-    Triple,
-    Unpaired,
-}
-
-// Filters all flops by the high card, ignoring suit
-pub fn high_flops(card: Card) -> Vec<[Card; 3]> {
-    high_flops_from_list(card, ALL_FLOPS.iter().copied().collect())
-}
-
-pub fn high_flops_from_list(card: Card, flop_list: Vec<[Card; 3]>) -> Vec<[Card; 3]> {
-    flop_list
-        .iter()
-        .copied()
-        .filter(|flop| flop[2] >> 2 == card >> 2)
-        .collect()
-}
-
-pub fn textured_flops(texture: Texture) -> Vec<[Card; 3]> {
-    textured_flops_from_list(texture, ALL_FLOPS.iter().copied().collect())
-}
-
-pub fn textured_flops_from_list(texture: Texture, flop_list: Vec<[Card; 3]>) -> Vec<[Card; 3]> {
-    flop_list
-        .iter()
-        .copied()
-        .filter(|flop| {
-            let suits: HashSet<Card> = HashSet::from_iter(flop.iter().copied().map(|c| c % 4));
-            let values: HashSet<Card> = HashSet::from_iter(flop.iter().copied().map(|c| c >> 2));
-            match texture {
-                Texture::Monotone => suits.len() == 1,
-                Texture::Twotone => suits.len() == 2,
-                Texture::Rainbow => suits.len() == 3,
-                Texture::Triple => values.len() == 1,
-                Texture::Paired => values.len() == 2,
-                Texture::Unpaired => values.len() == 3,
-            }
-        })
-        .collect()
-}
-
-pub fn flop_to_string(flop: &[u8; 3]) -> Result<String, String> {
-    Ok(flop
-        .iter()
-        .rev()
-        .map(|card| card_to_string(*card))
-        .collect::<Result<Vec<String>, String>>()?
-        .join(""))
-}
